set cwd for scripts
